# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/stats/advertisers/{advertiserId}/campaigns": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stats"
                ],
                "summary": "Get stats for all campaigns of this advertiser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "advertiserId",
                        "name": "advertiserId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CampaignStats"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    }
                }
            }
        },
        "/stats/advertisers/{advertiserId}/campaigns/daily": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stats"
                ],
                "summary": "Get daily stats for all campaigns of this advertiser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "advertiserId",
                        "name": "advertiserId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CampaignStats"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    }
                }
            }
        },
        "/stats/campaigns/{campaignId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stats"
                ],
                "summary": "Get stats for campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "campaignId",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CampaignStats"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    }
                }
            }
        },
        "/stats/campaigns/{campaignId}/daily": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stats"
                ],
                "summary": "Get daily stats for campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "campaignId",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CampaignStats"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    }
                }
            }
        },
...
        "/time/advance": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Time"
                ],
                "summary": "Update current date",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CurrentDate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CurrentDate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ginerr.ErrorResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "model.CampaignStats": {
            "type": "object",
            "properties": {
                "clicks_count": {
                    "type": "integer"
                },
                "conversion": {
                    "type": "number"
                },
                "date": {
                    "type": "integer"
                },
                "impressions_count": {
                    "type": "integer"
                },
                "spent_clicks": {
                    "type": "number"
                },
                "spent_impressions": {
                    "type": "number"
                },
                "spent_total": {
                    "type": "number"
                }
            }
        },
...
        "model.CurrentDate": {
            "type": "object",
            "required": [
                "current_date"
            ],
            "properties": {
                "current_date": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
    }
}